% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_classification.R
\name{assignClusterPhenotypes}
\alias{assignClusterPhenotypes}
\title{Assign Cluster Phenotypes}
\usage{
assignClusterPhenotypes(centers)
}
\description{
Assign Cluster Phenotypes
}
\keyword{internal}
