% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp_estimation.R
\name{atp_estimation}
\alias{atp_estimation}
\title{ATP Production Approximation Module}
\description{
Core novelty: Convert transcriptomics to quantitative ATP estimates
}
\keyword{ATP}
\keyword{energy}
\keyword{metabolism}
