% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{batchProcess}
\alias{batchProcess}
\title{Batch Process Large Dataset}
\usage{
batchProcess(seurat_obj, batch_size = 5000, FUN, ...)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{batch_size}{Number of cells per batch}

\item{FUN}{Function to apply to each batch}

\item{...}{Additional arguments to FUN}
}
\value{
Seurat object with results
}
\description{
Process large datasets in batches to manage memory.
}
\examples{
\dontrun{
obj <- batchProcess(seurat_obj, batch_size = 5000, myFunction, arg1 = val)
}
}
