% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{benchmarkFunction}
\alias{benchmarkFunction}
\title{Benchmark Function}
\usage{
benchmarkFunction(FUN, n_reps = 3, ...)
}
\arguments{
\item{FUN}{Function to benchmark}

\item{n_reps}{Number of repetitions}

\item{...}{Arguments to FUN}
}
\value{
List with benchmark results
}
\description{
Benchmark a function with timing and memory tracking.
}
