% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp_estimation.R
\name{calculateMetabolicPowerOutput}
\alias{calculateMetabolicPowerOutput}
\title{Calculate Metabolic Power Output}
\usage{
calculateMetabolicPowerOutput(
  seurat_obj,
  atp_col = "ATP_score",
  expression_rate = 1,
  assay = NULL,
  name = "metabolic_power"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{atp_col}{ATP score column}

\item{expression_rate}{Estimated expression rate}

\item{assay}{Assay to use for gene counts}

\item{name}{Column name}
}
\value{
Seurat object with power output
}
\description{
Calculate metabolic power as ATP per unit time, accounting
for the number of expressed metabolic genes.
}
