% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_integration.R
\name{calculateMetaboliteCorrelation}
\alias{calculateMetaboliteCorrelation}
\title{Calculate Metabolite Correlation}
\usage{
calculateMetaboliteCorrelation(
  seurat_obj,
  metabolite_data = NULL,
  score_cols = NULL
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{metabolite_data}{Metabolite abundance matrix}

\item{score_cols}{Score columns to correlate}
}
\value{
Matrix of correlations
}
\description{
Correlate metabolic scores with metabolite abundance.
}
\examples{
\dontrun{
metabolite_corr <- calculateMetaboliteCorrelation(seurat_obj, metabolite_data)
}
}
