% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_classification.R
\name{calculatePhenotypePurity}
\alias{calculatePhenotypePurity}
\title{Calculate Phenotype Purity Score}
\usage{
calculatePhenotypePurity(
  seurat_obj,
  phenotype_col = "metabolic_phenotype",
  cluster_col = "seurat_clusters"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{phenotype_col}{Phenotype column}

\item{cluster_col}{Cluster column}
}
\value{
Data frame with purity scores
}
\description{
Calculate purity of phenotype assignment within clusters.
}
\examples{
\dontrun{
purity <- calculatePhenotypePurity(seurat_obj)
print(purity)
}
}
