% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_aggregation.R
\name{celltype_aggregation}
\alias{celltype_aggregation}
\title{Cell-Type Level Aggregation Module}
\description{
Aggregate metabolic scores at cell-type level
}
\keyword{aggregation}
\keyword{cell}
\keyword{metabolism}
\keyword{type}
