% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{chunkApply}
\alias{chunkApply}
\title{Chunk Apply}
\usage{
chunkApply(data, chunk_size, FUN, n_cores = 1, ...)
}
\arguments{
\item{data}{Data to chunk}

\item{chunk_size}{Size of each chunk}

\item{FUN}{Function to apply}

\item{n_cores}{Number of cores}

\item{...}{Additional arguments to FUN}
}
\value{
List of results
}
\description{
Apply function to chunks of a vector/matrix in parallel.
}
