% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_classification.R
\name{classifyByHierarchical}
\alias{classifyByHierarchical}
\title{Classify by Hierarchical Clustering}
\usage{
classifyByHierarchical(glycolysis, oxphos, atp)
}
\description{
Classify by Hierarchical Clustering
}
\keyword{internal}
