% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_classification.R
\name{classifyByQuantile}
\alias{classifyByQuantile}
\title{Classify by Quantile Method}
\usage{
classifyByQuantile(
  glycolysis,
  oxphos,
  atp,
  glycolysis_threshold,
  oxphos_threshold,
  atp_low_threshold,
  atp_high_threshold
)
}
\description{
Classify by Quantile Method
}
\keyword{internal}
