% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_scoring.R
\name{computeAUCellScores}
\alias{computeAUCellScores}
\title{AUCell Scoring}
\usage{
computeAUCellScores(
  expr_matrix,
  gene_sets,
  nbin,
  nCores,
  verbose,
  genes_rownames
)
}
\description{
Internal function for AUCell-based scoring
}
\keyword{internal}
