% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_scoring.R
\name{computeGSVAScores}
\alias{computeGSVAScores}
\title{GSVA Scoring}
\usage{
computeGSVAScores(
  expr_matrix,
  gene_sets,
  k,
  tau,
  nCores,
  verbose,
  genes_rownames
)
}
\description{
Internal function for GSVA-based scoring
}
\keyword{internal}
