% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_scoring.R
\name{computeMeanScores}
\alias{computeMeanScores}
\title{Mean Expression Scoring}
\usage{
computeMeanScores(expr_matrix, gene_sets, verbose, genes_rownames)
}
\description{
Internal function for simple mean expression scoring
}
\keyword{internal}
