% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_scoring.R
\name{computeSingleSampleScores}
\alias{computeSingleSampleScores}
\title{Single Sample Scoring}
\usage{
computeSingleSampleScores(expr_matrix, gene_sets, verbose, genes_rownames)
}
\description{
Internal function for single sample gene set enrichment
}
\keyword{internal}
