% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_scoring.R
\name{computeWeightedMeanScores}
\alias{computeWeightedMeanScores}
\title{Weighted Mean Scoring}
\usage{
computeWeightedMeanScores(expr_matrix, gene_sets, verbose, genes_rownames)
}
\description{
Internal function for weighted mean expression scoring
}
\keyword{internal}
