% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{createProgressBar}
\alias{createProgressBar}
\title{Progress Bar Wrapper}
\usage{
createProgressBar(n, style = 3)
}
\arguments{
\item{n}{Total iterations}

\item{style}{Progress bar style}
}
\value{
Progress bar object
}
\description{
Create progress bar wrapper with timing.
}
