% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{createScMetaboFlux}
\alias{createScMetaboFlux}
\title{Create scMetaboFlux Object}
\usage{
createScMetaboFlux(seurat_obj, name = NULL)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{name}{Optional name for the object}
}
\value{
scMetaboFluxObject
}
\description{
Create an scMetaboFlux S4 object from Seurat for
structured storage and retrieval of results.
}
\examples{
\dontrun{
smf_obj <- createScMetaboFlux(seurat_obj, name = "my_analysis")
}
}
