% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_analysis.R
\name{differential_analysis}
\alias{differential_analysis}
\title{Differential Metabolism Analysis Module}
\description{
Compare metabolic states between conditions
}
\keyword{analysis}
\keyword{differential}
\keyword{metabolism}
