% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_approximation.R
\name{estimateFluxDirection}
\alias{estimateFluxDirection}
\title{Estimate Metabolic Flux Direction}
\usage{
estimateFluxDirection(
  seurat_obj,
  glycolysis_col = "glycolysis_flux",
  oxphos_col = "oxidative_phosphorylation_flux",
  name = "flux_direction"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{glycolysis_col}{Glycolysis flux column}

\item{oxphos_col}{OXPHOS flux column}

\item{name}{Column name}
}
\value{
Seurat object with flux direction
}
\description{
Estimate predominant flux direction (glycolytic vs oxidative).
}
