% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp_estimation.R
\name{estimateMetabolicEfficiency}
\alias{estimateMetabolicEfficiency}
\title{Estimate Metabolic Efficiency}
\usage{
estimateMetabolicEfficiency(
  seurat_obj,
  atp_col = "ATP_score",
  glycolysis_col = "glycolysis",
  oxphos_col = "oxidative_phosphorylation",
  name = "metabolic_efficiency"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{atp_col}{ATP score column}

\item{glycolysis_col}{Glycolysis score column}

\item{oxphos_col}{OXPHOS score column}

\item{name}{Column name for efficiency}
}
\value{
Seurat object with efficiency added
}
\description{
Calculate metabolic efficiency index as ATP per unit of
total pathway activity.
}
