% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{exportMetabolicResults}
\alias{exportMetabolicResults}
\title{Export Results to File}
\usage{
exportMetabolicResults(
  seurat_obj,
  output_dir = "scMetaboFlux_results",
  include_embeddings = TRUE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{output_dir}{Output directory}

\item{include_embeddings}{Include UMAP/tSNE embeddings}
}
\description{
Export metabolic analysis results to CSV files.
}
\examples{
\dontrun{
exportMetabolicResults(seurat_obj, 
  output_dir = "scMetaboFlux_results",
  include_embeddings = TRUE)
}
}
