% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{extractResults}
\alias{extractResults}
\alias{extractResults,scMetaboFluxObject-method}
\title{Extract Results from scMetaboFlux}
\usage{
extractResults(object, what)

\S4method{extractResults}{scMetaboFluxObject}(object, what = c("scores", "phenotypes", "flux", "all"))
}
\arguments{
\item{object}{scMetaboFlux object}

\item{what}{What to extract: "scores", "phenotypes", "flux", "all"}
}
\value{
Requested data
}
\description{
Extract specific results from scMetaboFlux object.
}
\examples{
\dontrun{
smf_obj <- createScMetaboFlux(seurat_obj)
scores <- extractResults(smf_obj, what = "scores")
}
}
