% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_approximation.R
\name{flux_approximation}
\alias{flux_approximation}
\title{Metabolic Flux Approximation Module}
\description{
Compute pseudo-flux scores for metabolic pathways
}
\keyword{flux}
\keyword{metabolic}
\keyword{modeling}
