% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{getAnalysisParameters}
\alias{getAnalysisParameters}
\title{Get Analysis Parameters}
\usage{
getAnalysisParameters(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat object}
}
\value{
List of analysis parameters or NULL
}
\description{
Get parameters used for analysis.
}
\examples{
\dontrun{
params <- getAnalysisParameters(seurat_obj)
}
}
