% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_aggregation.R
\name{getDominantPhenotypePerCellType}
\alias{getDominantPhenotypePerCellType}
\title{Get Dominant Phenotype per Cell Type}
\usage{
getDominantPhenotypePerCellType(
  seurat_obj,
  cell_type_col = "cell_type",
  phenotype_col = "metabolic_phenotype"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{cell_type_col}{Cell type column}

\item{phenotype_col}{Phenotype column}
}
\value{
Data frame with dominant phenotypes
}
\description{
Get dominant metabolic phenotype for each cell type.
}
\examples{
\dontrun{
dominant <- getDominantPhenotypePerCellType(seurat_obj, cell_type_col = "cell_type")
print(dominant)
}
}
