% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{getOptimalBatchSize}
\alias{getOptimalBatchSize}
\title{Get Optimal Batch Size}
\usage{
getOptimalBatchSize(seurat_obj, target_memory_pct = 50)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{target_memory_pct}{Target memory usage percentage}
}
\value{
Recommended batch size
}
\description{
Calculate optimal batch size based on available memory.
}
