% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{hasMetabolicScores}
\alias{hasMetabolicScores}
\title{Check if Scores Exist}
\usage{
hasMetabolicScores(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat object}
}
\value{
Logical indicating if scores exist
}
\description{
Check if metabolic scores have been computed.
}
\examples{
\dontrun{
hasMetabolicScores(seurat_obj)
}
}
