% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{initialize,scMetaboFluxObject-method}
\alias{initialize,scMetaboFluxObject-method}
\title{Initialize scMetaboFlux Object}
\usage{
\S4method{initialize}{scMetaboFluxObject}(.Object, ...)
}
\arguments{
\item{.Object}{A scMetaboFlux object}

\item{...}{Additional arguments}
}
\description{
Initialize scMetaboFlux Object
}
