% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_integration.R
\name{integrateProteomicsData}
\alias{integrateProteomicsData}
\title{Integrate Proteomics Data}
\usage{
integrateProteomicsData(
  seurat_obj,
  protein_data = NULL,
  metabolic_proteins = NULL
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{protein_data}{Protein expression matrix (cells x proteins)}

\item{metabolic_proteins}{Known metabolic proteins to analyze}
}
\value{
Data frame with correlations
}
\description{
Correlate metabolic gene expression with protein abundance.
}
\examples{
\dontrun{
protein_corr <- integrateProteomicsData(seurat_obj)
}
}
