% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{mergeMetabolicGeneSets}
\alias{mergeMetabolicGeneSets}
\title{Merge Metabolic Gene Sets}
\usage{
mergeMetabolicGeneSets(
  gene_sets,
  remove_duplicates = TRUE,
  case_sensitive = FALSE
)
}
\arguments{
\item{gene_sets}{List of gene set character vectors to merge}

\item{remove_duplicates}{Remove duplicate genes (default TRUE)}

\item{case_sensitive}{Whether duplicate detection should be case-sensitive}
}
\value{
Character vector of merged genes
}
\description{
Merges multiple gene sets into a single character vector,
optionally removing duplicate genes.
}
\examples{
set1 <- c("HK2", "PKM2", "LDHA")
set2 <- c("LDHA", "NDUFA9", "SDHA")
merged <- mergeMetabolicGeneSets(list(set1, set2))
merged
}
