% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\name{metabolic_gene_mapping}
\alias{metabolic_gene_mapping}
\title{Metabolic Gene Mapping Engine}
\description{
Maps genes to metabolic pathways using KEGG/Reactome databases
}
\keyword{gene}
\keyword{mapping}
\keyword{metabolic}
\keyword{pathways}
