% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\docType{data}
\name{metabolicGeneSets}
\alias{metabolicGeneSets}
\title{Predefined Metabolic Gene Sets}
\format{
A named list where each element is a character vector of gene symbols:
\describe{
\item{glycolysis}{37 genes involved in glucose breakdown to pyruvate}
\item{tca_cycle}{40 genes for citric acid cycle enzymes}
\item{oxidative_phosphorylation}{150+ genes for ETC complexes I-V}
\item{fatty_acid_oxidation}{45 genes for beta-oxidation pathway}
\item{pentose_phosphate_pathway}{17 genes for PPP enzymes}
\item{glutaminolysis}{11 genes for glutamine metabolism}
\item{pyruvate_metabolism}{14 genes for pyruvate conversion}
\item{electron_transport_chain}{90+ genes for ETC}
\item{mitochondrial_biogenesis}{30 genes for mtDNA replication}
\item{hypoxia_response}{24 genes for hypoxic response}
\item{glycolysis_rate_limiting}{12 key rate-limiting enzymes}
\item{oxidative_phosphorylation_rate_limiting}{8 key rate-limiting enzymes}
\item{tca_rate_limiting}{8 key rate-limiting enzymes}
}
}
\source{
KEGG Pathway Database (https://www.kegg.jp/kegg/pathway.html)

Reactome Database (https://reactome.org/)
}
\usage{
metabolicGeneSets
}
\description{
Comprehensive metabolic gene sets curated from KEGG and Reactome databases.
Contains 12 metabolic pathways with validated gene memberships for human and mouse.
}
\examples{
head(metabolicGeneSets$glycolysis)
length(metabolicGeneSets$oxidative_phosphorylation)

}
\keyword{datasets}
