% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiomics_integration.R
\name{multiomics_integration}
\alias{multiomics_integration}
\title{Multi-Omics Integration Module}
\description{
Integrate scMetaboFlux with other omics layers
}
\keyword{integration}
\keyword{multiomics}
\keyword{proteomics}
\keyword{spatial}
