% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_optimization.R
\name{optimizeSeuratObject}
\alias{optimizeSeuratObject}
\title{Optimize Seurat Object}
\usage{
optimizeSeuratObject(seurat_obj, remove_scale_data = TRUE, compress = TRUE)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{remove_scale_data}{Remove scale.data to save memory}

\item{compress}{Compress assay data}
}
\value{
Optimized Seurat object
}
\description{
Optimize Seurat object for memory efficiency.
}
