% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_scoring.R
\name{pathway_scoring}
\alias{pathway_scoring}
\title{Pathway Activity Scoring Module}
\description{
Compute pathway activity scores using multiple methods
}
\keyword{enrichment}
\keyword{gene}
\keyword{pathway}
\keyword{scoring}
\keyword{set}
