% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_classification.R
\name{phenotype_classification}
\alias{phenotype_classification}
\title{Metabolic Phenotype Classification Module}
\description{
Classify cells into metabolic phenotypes
}
\keyword{classification}
\keyword{metabolism}
\keyword{phenotype}
