% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotATPDistribution}
\alias{plotATPDistribution}
\title{Plot ATP Distribution}
\usage{
plotATPDistribution(
  seurat_obj,
  atp_col = "ATP_score",
  group_col = NULL,
  type = c("violin", "boxplot", "histogram", "density")
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{atp_col}{ATP score column}

\item{group_col}{Grouping variable}

\item{type}{Plot type: "violin", "boxplot", "histogram", "density"}
}
\value{
ggplot
}
\description{
Plot ATP score distribution.
}
\examples{
\dontrun{
plotATPDistribution(seurat_obj, group_col = "cell_type", type = "violin")
}
}
