% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCellTypeProfiles}
\alias{plotCellTypeProfiles}
\title{Plot Cell Type Metabolic Profiles}
\usage{
plotCellTypeProfiles(
  seurat_obj,
  cell_type_col = "cell_type",
  score_cols = NULL,
  top_n = 10
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{cell_type_col}{Cell type column}

\item{score_cols}{Score columns}

\item{top_n}{Number of top cell types to show}
}
\value{
ggplot (bar chart)
}
\description{
Plot metabolic profiles by cell type.
}
\examples{
\dontrun{
plotCellTypeProfiles(seurat_obj, cell_type_col = "cell_type")
}
}
