% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotGlycolysisVsOxphos}
\alias{plotGlycolysisVsOxphos}
\title{Plot Glycolysis vs OXPHOS}
\usage{
plotGlycolysisVsOxphos(
  seurat_obj,
  glycolysis_col = "glycolysis",
  oxphos_col = "oxidative_phosphorylation",
  color_by = NULL,
  density = TRUE,
  sample_n = NULL
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{glycolysis_col}{Glycolysis score column}

\item{oxphos_col}{OXPHOS score column}

\item{color_by}{Color points by this variable}

\item{density}{Add density contours}

\item{sample_n}{Sample number of points for large datasets}
}
\value{
ggplot
}
\description{
Create scatter plot of glycolysis vs OXPHOS.
}
\examples{
\dontrun{
plotGlycolysisVsOxphos(seurat_obj, color_by = "cell_type")
}
}
