% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotGOXIndex}
\alias{plotGOXIndex}
\title{Plot GOX Index Distribution}
\usage{
plotGOXIndex(seurat_obj, gox_col = "GOX_index", group_col = NULL)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{gox_col}{GOX index column}

\item{group_col}{Grouping variable}
}
\value{
ggplot
}
\description{
Plot GOX (Glycolytic-OXPHOS) index distribution.
}
\examples{
\dontrun{
plotGOXIndex(seurat_obj, group_col = "cell_type")
}
}
