% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{resetMetabolicResults}
\alias{resetMetabolicResults}
\title{Reset Metabolic Results}
\usage{
resetMetabolicResults(seurat_obj, keep_original = FALSE)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{keep_original}{Keep original scores if present}
}
\value{
Seurat object with results removed
}
\description{
Remove all scMetaboFlux results from Seurat object.
}
\examples{
\dontrun{
seurat_obj <- resetMetabolicResults(seurat_obj)
}
}
