% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_workflow.R
\name{summary,scMetaboFluxObject-method}
\alias{summary,scMetaboFluxObject-method}
\title{Summary Method for scMetaboFlux}
\usage{
\S4method{summary}{scMetaboFluxObject}(object)
}
\arguments{
\item{object}{A scMetaboFlux object}
}
\description{
Summary Method for scMetaboFlux
}
